backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.2,<4.0
